+++ src/routes/ipobj/types.js
@@ -23,12 +23,12 @@
 
 var express = require('express');
 var router = express.Router();
-var Ipobj_typeModel = require('../../models/ipobj/ipobj_type');
+import { IPObjType } from '../../models/ipobj/IPObjType';
 const fwcError = require('../../utils/error_table');
 
 /* Get all ipobj_types */
 router.get('/', (req, res) => {
-	Ipobj_typeModel.getIpobj_types((error, data) => {
+	IPObjType.getIpobj_types((error, data) => {
 		if (error) return res.status(400).json(error);
 
 		if (data && data.length > 0)
@@ -41,7 +41,7 @@ router.get('/', (req, res) => {
 /* Get  ipobj_type by id */
 router.put('/get', async (req, res) => {
 	try {
-		const data = await Ipobj_typeModel.getIpobj_type(req, req.body.id);		
+		const data = await IPObjType.getIpobj_type(req, req.body.id);		
     if (data && data.length > 0)
       res.status(200).json(data);
     else
